requirements.txt
Django>=3.0,<4.0
djangorestframework
python-decouple
python-dotenv
mysqlclient
django-cors-headers
pynfe
